+++ aimlflow/cli.py
@@ -25,7 +25,9 @@ def cli_entry_point():
                                                             writable=True))
 @click.option('--mlflow-tracking-uri',  required=False, default=None)
 @click.option('--experiment', '-e', required=False, default=None)
-def sync(aim_repo, mlflow_tracking_uri, experiment):
+@click.option('--exclude-artifacts', multiple=True, required=False)
+def sync(aim_repo, mlflow_tracking_uri, experiment, exclude_artifacts):
+
     repo_path = clean_repo_path(aim_repo) or Repo.default_repo_path()
     repo_inst = Repo.from_path(repo_path)
 
@@ -33,10 +35,10 @@ def sync(aim_repo, mlflow_tracking_uri, experiment):
     if not mlflow_tracking_uri:
         raise ClickException('MLFlow tracking URI must be provided either through ENV or CLI.')
 
-    watcher = MLFlowWatcher(repo_inst, mlflow_tracking_uri, experiment)
+    watcher = MLFlowWatcher(repo_inst, mlflow_tracking_uri, experiment, exclude_artifacts)
 
     click.echo('Converting existing MLflow logs.')
-    convert_existing_logs(repo_inst, mlflow_tracking_uri, experiment)
+    convert_existing_logs(repo_inst, mlflow_tracking_uri, experiment, exclude_artifacts)
 
     click.echo(f'Starting watcher on {mlflow_tracking_uri}.')
     watcher.start()