+++ aimlflow/utils.py
@@ -1,3 +1,4 @@
+import fnmatch
 import click
 import collections
 import mlflow
@@ -117,7 +118,10 @@ def collect_run_params(aim_run, mlflow_run):
     }
 
 
-def collect_artifacts(aim_run, mlflow_run, mlflow_client):
+def collect_artifacts(aim_run, mlflow_run, mlflow_client, exclude_artifacts):
+    if '*' in exclude_artifacts:
+        return
+
     run_id = mlflow_run.info.run_id
 
     artifacts_cache_key = '_mlflow_artifacts_cache'
@@ -139,6 +143,16 @@ def collect_artifacts(aim_run, mlflow_run, mlflow_client):
                     continue
                 else:
                     artifacts_cache.append(file_info.path)
+
+                if exclude_artifacts:
+                    exclude = False
+                    for expr in exclude_artifacts:
+                        if fnmatch.fnmatch(file_info.path, expr):
+                            exclude = True
+                            break
+                    if exclude:
+                        continue
+
                 downloaded_path = mlflow_client.download_artifacts(run_id, file_info.path, dst_path=temp_path)
                 if file_info.path.endswith(HTML_EXTENSIONS):
                     if not __html_warning_issued:
@@ -199,7 +213,7 @@ def collect_metrics(aim_run, mlflow_run, mlflow_client, timestamp=None):
             aim_run.track(m.value, step=m.step, name=m.key)
 
 
-def convert_existing_logs(repo_inst, tracking_uri, experiment=None, no_cache=False):
+def convert_existing_logs(repo_inst, tracking_uri, experiment=None, excluded_artifacts=None, no_cache=False):
     client = mlflow.tracking.client.MlflowClient(tracking_uri=tracking_uri)
 
     experiments = get_mlflow_experiments(client, experiment)
@@ -217,7 +231,7 @@ def convert_existing_logs(repo_inst, tracking_uri, experiment=None, no_cache=Fal
             collect_metrics(aim_run, run, client)
 
             # Collect artifacts
-            collect_artifacts(aim_run, run, client)
+            collect_artifacts(aim_run, run, client, excluded_artifacts)
 
     run_cache.refresh()